pyproject.toml
[project]
name = "repomap-tool"
description = "Standalone repository map generator extracted from the aider project"
readme = "README.md"
license = { file = "../LICENSE.txt" }
authors = [{ name = "Aider Developers" }]
requires-python = ">=3.10"
version = "0.1.0"
dependencies = [
    "diskcache>=5.6",
    "GitPython>=3.1",
    "grep-ast>=0.4",
    "networkx<3.5",
    "scipy<1.16",
    "pygments>=2.16",
    "tqdm>=4.64",
    "tree-sitter-languages>=1.10.2",
]

[project.urls]
Homepage = "https://github.com/Aider-AI/aider/tree/main/repomap_tool"

[project.scripts]
repomap-tool = "repomap_tool.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["repomap_tool", "aider"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
